+++ src/components/RotaryDial.tsx
@@ -1,50 +1,5 @@
 import React, { useState, useRef, useCallback, useEffect } from "react";
-import styled from "styled-components";
-
-interface DisplayProps {
-  rotation: number;
-}
-
-const DialContainer = styled.div`
-  position: relative;
-  width: 200px;
-  height: 200px;
-`;
-
-const Dial = styled.div<{ rotation: number }>`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  border: 8px solid #aaa;
-  border-radius: 50%;
-  background: #f3f3f3;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transform: ${({ rotation }) => `rotate(${rotation}deg)`};
-`;
-
-const Display = styled.div<DisplayProps>`
-  font-size: 24px;
-  font-weight: bold;
-  z-index: 1;
-  pointer-events: none;
-  transform: ${({ rotation }) => `rotate(${-rotation}deg)`};
-  user-select: none;
-`;
-
-const Handle = styled.div`
-  width: 30px;
-  height: 30px;
-  background: #333;
-  border-radius: 50%;
-  position: absolute;
-  top: -15px;
-  left: calc(50% - 15px);
-  cursor: grab;
-`;
+import * as S from "./RotaryDial.styles";
 
 const RotaryDial = () => {
   const [rotation, setRotation] = useState(0);
@@ -108,12 +63,12 @@ const RotaryDial = () => {
   }, [handleMouseMove, handleMouseUp]);
 
   return (
-    <DialContainer>
-      <Dial className="dial" ref={dialRef} rotation={rotation} onMouseDown={handleMouseDown}>
-        <Display rotation={rotation}>{selectedNumber}</Display>
-        <Handle />
-      </Dial>
-    </DialContainer>
+    <S.DialContainer>
+      <S.Dial className="dial" ref={dialRef} rotation={rotation} onMouseDown={handleMouseDown}>
+        <S.Display rotation={rotation}>{selectedNumber}</S.Display>
+        <S.Handle />
+      </S.Dial>
+    </S.DialContainer>
   );
 };
